user_memの型をintptr_t[]に変更

+++ src/global.cpp
@@ -8,7 +8,7 @@ body **rstack;
 intptr_t stack_start[FORTH_STACK_SIZE];
 body *rstack_start[FORTH_RSTACK_SIZE];
 char input_buf[FORTH_INPUT_BUF];
-int32_t user_mem[FORTH_USER_MEM];
+intptr_t user_mem[FORTH_USER_MEM];
 forth_word dict_mem[FORTH_DICT_MEM];
 
 uint32_t state = 0;
