+++ src/global.hpp
@@ -19,7 +19,7 @@ extern body **rstack;
 extern intptr_t stack_start[FORTH_STACK_SIZE];
 extern body *rstack_start[FORTH_RSTACK_SIZE];
 extern char input_buf[FORTH_INPUT_BUF];
-extern int32_t user_mem[FORTH_USER_MEM];
+extern intptr_t user_mem[FORTH_USER_MEM];
 extern forth_word dict_mem[FORTH_DICT_MEM];
 
 extern uint32_t state;
